ethwallet transfer and listcoin

+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/mapper/UserWalletMapper.java
@@ -1,6 +1,6 @@
 package com.cascv.oas.server.blockchain.mapper;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import org.apache.ibatis.annotations.Param;
 
@@ -10,8 +10,8 @@ public interface UserWalletMapper {
   Integer insertSelective(UserWallet userWallet);
   UserWallet selectByUuid(String uuid);
   UserWallet selectByUserUuid(String userUuid);
-  Integer increaseBalance(@Param("uuid")  String uuid, @Param("value")  BigInteger value);
-  Integer decreaseBalance(@Param("uuid")  String uuid, @Param("value")  BigInteger value);
+  Integer increaseBalance(@Param("uuid")  String uuid, @Param("value")  BigDecimal value);
+  Integer decreaseBalance(@Param("uuid")  String uuid, @Param("value")  BigDecimal value);
   
   Integer deleteByUuid(String uuid);
   Integer deleteByUserUuid(String userUuid);
