+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/model/UserWallet.java
@@ -1,7 +1,7 @@
 package com.cascv.oas.server.blockchain.model;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -10,7 +10,7 @@ public class UserWallet implements Serializable{
   private static final long serialVersionUID = 1L;
   @Getter @Setter private String uuid;
   @Getter @Setter private String userUuid;
-  @Getter @Setter private BigInteger balance;
+  @Getter @Setter private BigDecimal balance;
   @Getter @Setter private String created;
   @Getter @Setter private String updated;
 }
