+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/vo/UserWalletTransfer.java
@@ -1,7 +1,7 @@
 package com.cascv.oas.server.blockchain.vo;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,5 +9,5 @@ import lombok.Setter;
 public class UserWalletTransfer implements Serializable {
   private static final long serialVersionUID = 1L;
 	@Getter @Setter private String toUserName;
-  @Getter @Setter private BigInteger value;
+  @Getter @Setter private BigDecimal value;
 }
